refactor(app-page-builder): move `useConfigureDomain` to hooks
wrap `handlePreviewClick` in `useCallback` hook
