+++ packages/app-page-builder/src/admin/hooks/useConfigureDomain/index.ts
@@ -1,0 +1,1 @@
+export { useConfigureDomainDialog } from "./useConfigureDomain";
